+++ src/ft_math/vec2_rotate.c
@@ -6,17 +6,17 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 12:52:24 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/11/10 13:01:22 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/11/11 12:40:28 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <math.h>
 
-t_vec2	vec2_rotate(t_vec2 v, float angle)
+t_vec2	vec2_rotate(t_vec2 v, double angle)
 {
-	float	len;
+	double	len;
 
 	len = vec2_length(v);
-	return ((t_vec2){.x = cosf(angle) * len, .y = sinf(angle) * len});
+	return ((t_vec2){.x = cos(angle) * len, .y = sin(angle) * len});
 }
